examples/llama_bench_performance/requirement.txt
pyaml==25.5.0